core: proxy exports
