+++ packages/core/src/exports/constants.ts
@@ -1,0 +1,5 @@
+////////////////////////////////////////////////////////////////////////////////
+// Constants
+////////////////////////////////////////////////////////////////////////////////
+
+export * from '../constants.js'
