+++ packages/core/src/exports/query.ts
@@ -1,0 +1,50 @@
+////////////////////////////////////////////////////////////////////////////////
+// Tanstack Query
+////////////////////////////////////////////////////////////////////////////////
+
+export {
+  type GetBackOfQueueWalletData,
+  type GetBackOfQueueWalletOptions,
+  type GetBackOfQueueWalletQueryFnData,
+  type GetBackOfQueueWalletQueryKey,
+  getBackOfQueueWalletQueryKey,
+  getBackOfQueueWalletQueryOptions,
+} from '../query/getBackOfQueueWallet.js'
+
+export {
+  type GetOrderHistoryData,
+  type GetOrderHistoryOptions,
+  type GetOrderHistoryQueryFnData,
+  type GetOrderHistoryQueryKey,
+  getOrderHistoryQueryKey,
+  getOrderHistoryQueryOptions,
+} from '../query/getOrderHistory.js'
+
+export {
+  type GetNetworkOrdersData,
+  type GetNetworkOrdersOptions,
+  type GetNetworkOrdersQueryFnData,
+  type GetNetworkOrdersQueryKey,
+  getNetworkOrdersQueryKey,
+  getNetworkOrdersQueryOptions,
+} from '../query/getNetworkOrders.js'
+
+export {
+  type GetTaskHistoryData,
+  type GetTaskHistoryOptions,
+  type GetTaskHistoryQueryFnData,
+  type GetTaskHistoryQueryKey,
+  getTaskHistoryQueryKey,
+  getTaskHistoryQueryOptions,
+} from '../query/getTaskHistory.js'
+
+export {
+  type GetWalletData,
+  type GetWalletOptions,
+  type GetWalletQueryFnData,
+  type GetWalletQueryKey,
+  getWalletQueryKey,
+  getWalletQueryOptions,
+} from '../query/getWallet.js'
+
+export { hashFn, structuralSharing } from '../query/utils.js'
