--- packages/core/src/index.ts
@@ -1,275 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-// createConfig
-////////////////////////////////////////////////////////////////////////////////
-
-export {
-  type Config,
-  type CreateConfigParameters,
-  type State,
-  createConfig,
-} from './createConfig.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Actions
-////////////////////////////////////////////////////////////////////////////////
-
-export {
-  type CancelOrderParameters,
-  type CancelOrderReturnType,
-  cancelOrder,
-} from './actions/cancelOrder.js'
-
-export {
-  type ConnectParameters,
-  type ConnectReturnType,
-  connect,
-} from './actions/connect.js'
-
-export {
-  type CreateOrderParameters,
-  type CreateOrderReturnType,
-  createOrder,
-} from './actions/createOrder.js'
-
-export {
-  type CreateWalletReturnType,
-  createWallet,
-} from './actions/createWallet.js'
-
-export {
-  type DepositParameters,
-  type DepositReturnType,
-  deposit,
-} from './actions/deposit.js'
-
-export {
-  type DisconnectReturnType,
-  disconnect,
-} from './actions/disconnect.js'
-
-export {
-  type GetAuthorizationHeadersParameters,
-  type GetAuthorizationHeadersReturnType,
-  getAuthorizationHeaders,
-} from './actions/getAuthHeaders.js'
-
-export {
-  type GetBalancesReturnType,
-  getBalances,
-} from './actions/getBalances.js'
-
-export {
-  type GetBackOfQueueWalletParameters,
-  type GetBackOfQueueWalletReturnType,
-  getBackOfQueueWallet,
-} from './actions/getBackOfQueueWallet.js'
-
-export {
-  type GetNetworkOrdersReturnType,
-  type GetNetworkOrdersErrorType,
-  getNetworkOrders,
-} from './actions/getNetworkOrders.js'
-
-export {
-  type GetOrderParameters,
-  type GetOrderReturnType,
-  getOrder,
-} from './actions/getOrder.js'
-
-export {
-  type GetOrderHistoryErrorType,
-  type GetOrderHistoryReturnType,
-  getOrderHistory,
-} from './actions/getOrderHistory.js'
-
-export {
-  type GetOrdersReturnType,
-  getOrders,
-} from './actions/getOrders.js'
-
-export {
-  type GetPkRootReturnType,
-  getPkRoot,
-  type GetPkRootScalarsReturnType,
-  getPkRootScalars,
-} from './actions/getPkRoot.js'
-
-export {
-  type GetPriceParameters,
-  type GetPriceReturnType,
-  getPriceFromPriceReporter,
-} from './actions/getPriceFromPriceReporter.js'
-
-export {
-  type GetSkRootParameters,
-  type GetSkRootReturnType,
-  getSkRoot,
-} from './actions/getSkRoot.js'
-
-export {
-  type GetTaskHistoryErrorType,
-  type GetTaskHistoryReturnType,
-  getTaskHistory,
-} from './actions/getTaskHistory.js'
-
-export {
-  type GetTaskStatusParameters,
-  type GetTaskStatusReturnType,
-  getTaskStatus,
-} from './actions/getTaskStatus.js'
-
-export {
-  type GetTaskQueueReturnType,
-  getTaskQueue,
-} from './actions/getTaskQueue.js'
-
-export {
-  type GetWalletFromRelayerParameters,
-  type GetWalletFromRelayerReturnType,
-  type GetWalletFromRelayerErrorType as GetWalletErrorType,
-  getWalletFromRelayer,
-} from './actions/getWalletFromRelayer.js'
-
-export {
-  type GetWalletIdParameters,
-  type GetWalletIdReturnType,
-  getWalletId,
-} from './actions/getWalletId.js'
-
-export {
-  type LookupWalletParameters,
-  type LookupWalletReturnType,
-  lookupWallet,
-} from './actions/lookupWallet.js'
-
-export { type PayFeesReturnType, payFees } from './actions/payFees.js'
-
-export {
-  type SignMessageParameters,
-  type SignMessageReturnType,
-  signMessage,
-} from './actions/signMessage.js'
-
-export {
-  type UpdateOrderParameters,
-  type UpdateOrderReturnType,
-  updateOrder,
-} from './actions/updateOrder.js'
-
-export {
-  type WaitForTaskCompletionParameters,
-  type WaitForTaskCompletionReturnType,
-  waitForTaskCompletion,
-} from './actions/waitForTaskCompletion.js'
-
-export {
-  type WatchStatusParameters,
-  type WatchStatusReturnType,
-  watchStatus,
-} from './actions/watchStatus.js'
-
-export {
-  type WithdrawParameters,
-  type WithdrawReturnType,
-  withdraw,
-} from './actions/withdraw.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Hydrate
-////////////////////////////////////////////////////////////////////////////////
-
-export { hydrate } from './hydrate.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Errors
-////////////////////////////////////////////////////////////////////////////////
-
-export { BaseError } from './errors/base.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Constants
-////////////////////////////////////////////////////////////////////////////////
-export * from './constants.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Types
-////////////////////////////////////////////////////////////////////////////////
-
-export { Token } from './types/token.js'
-
-export * from './types/wallet.js'
-
-export * from './types/utils.js'
-
-export type { Task, TaskInfo, TaskState } from './types/task.js'
-export { UpdateType, TaskType } from './types/task.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Utils
-////////////////////////////////////////////////////////////////////////////////
-
-export { formatAmount, parseAmount } from './utils/format.js'
-
-export { deepEqual } from './utils/deepEqual.js'
-
-export { postRelayerRaw } from './utils/http.js'
-
-export { WebSocketManager } from './utils/websocket.js'
-
-export { parseBigJSON } from './utils/bigJSON.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Viem
-////////////////////////////////////////////////////////////////////////////////
-export { chain } from './utils/viem.js'
-
-////////////////////////////////////////////////////////////////////////////////
-// Tanstack Query
-////////////////////////////////////////////////////////////////////////////////
-
-export { hashFn, structuralSharing } from './query/utils.js'
-
-export {
-  type GetBackOfQueueWalletData,
-  type GetBackOfQueueWalletOptions,
-  type GetBackOfQueueWalletQueryFnData,
-  type GetBackOfQueueWalletQueryKey,
-  getBackOfQueueWalletQueryKey,
-  getBackOfQueueWalletQueryOptions,
-} from './query/getBackOfQueueWallet.js'
-
-export {
-  type GetWalletData,
-  type GetWalletOptions,
-  type GetWalletQueryFnData,
-  type GetWalletQueryKey,
-  getWalletQueryKey,
-  getWalletQueryOptions,
-} from './query/getWallet.js'
-
-export {
-  type GetTaskHistoryData,
-  type GetTaskHistoryOptions,
-  type GetTaskHistoryQueryFnData,
-  type GetTaskHistoryQueryKey,
-  getTaskHistoryQueryKey,
-  getTaskHistoryQueryOptions,
-} from './query/getTaskHistory.js'
-
-export {
-  type GetOrderHistoryData,
-  type GetOrderHistoryOptions,
-  type GetOrderHistoryQueryFnData,
-  type GetOrderHistoryQueryKey,
-  getOrderHistoryQueryKey,
-  getOrderHistoryQueryOptions,
-} from './query/getOrderHistory.js'
-
-export {
-  type GetNetworkOrdersData,
-  type GetNetworkOrdersOptions,
-  type GetNetworkOrdersQueryFnData,
-  type GetNetworkOrdersQueryKey,
-  getNetworkOrdersQueryKey,
-  getNetworkOrdersQueryOptions,
-} from './query/getNetworkOrders.js'
