+++ packages/node/src/index.ts
@@ -1,5 +1,5 @@
 export * from '@renegade-fi/core'
-import { createConfig as core_createConfig } from '@renegade-fi/core'
+import { createConfig as core_createConfig } from '@renegade-fi/core/'
 
 import * as RustUtils from '../renegade-utils/index.js'
