+++ packages/react/src/hooks/useBackOfQueueWallet.ts
@@ -1,20 +1,19 @@
 'use client'
 
+import type { Evaluate, GetBackOfQueueWalletErrorType } from '@renegade-fi/core'
 import {
   getBackOfQueueWalletQueryOptions,
-  type Evaluate,
   type GetBackOfQueueWalletData,
   type GetBackOfQueueWalletOptions,
   type GetBackOfQueueWalletQueryFnData,
   type GetBackOfQueueWalletQueryKey,
-} from '@renegade-fi/core'
+} from '@renegade-fi/core/query'
 import { useQueryClient } from '@tanstack/react-query'
 import type { ConfigParameter, QueryParameter } from '../types/properties.js'
 import { useQuery, type UseQueryReturnType } from '../utils/query.js'
 import { useConfig } from './useConfig.js'
 import { useStatus } from './useStatus.js'
 import { useWalletWebsocket } from './useWalletWebSocket.js'
-import type { GetBackOfQueueWalletErrorType } from '@renegade-fi/core/dist/types/actions/getBackOfQueueWallet.js'
 
 export type UseBackOfQueueWalletParameters<
   selectData = GetBackOfQueueWalletData,
