+++ packages/react/src/utils/query.ts
@@ -1,4 +1,9 @@
+import type { Evaluate, ExactPartial, Omit, UnionOmit } from '@renegade-fi/core'
+import { hashFn } from '@renegade-fi/core/query'
 import {
+  useInfiniteQuery as tanstack_useInfiniteQuery,
+  useQuery as tanstack_useQuery,
+  useMutation,
   type DefaultError,
   type QueryKey,
   type UseInfiniteQueryOptions,
@@ -7,17 +12,7 @@ import {
   type UseMutationResult,
   type UseQueryOptions,
   type UseQueryResult,
-  useInfiniteQuery as tanstack_useInfiniteQuery,
-  useMutation,
-  useQuery as tanstack_useQuery,
 } from '@tanstack/react-query'
-import {
-  type Evaluate,
-  type ExactPartial,
-  type Omit,
-  type UnionOmit,
-} from '@renegade-fi/core'
-import { hashFn } from '@renegade-fi/core'
 
 export type UseMutationParameters<
   data = unknown,
